+++ pysat/__init__.py
@@ -60,16 +60,16 @@ if not os.path.isdir(pysat_dir):
     os.mkdir(pysat_dir)
     print('Created .pysat directory in user home directory to store settings.')
     # create file with default data directory
-    if not (os.environ.get('TRAVIS') == 'true'):
-        data_dir = os.path.join(home_dir, 'pysatData')
-    else:
+    if (os.environ.get('TRAVIS') == 'true'):
         data_dir = '/home/travis/build/pysatData'
+    else:
+        data_dir = ''
     with open(os.path.join(pysat_dir, 'data_path.txt'), 'w') as f:
         f.write(data_dir)
-    print(''.join(("\nHi there!  Pysat will nominally store data in the "
-                   "'pysatData' directory at the user's home directory level. "
-                   "Run pysat.utils.set_data_dir to specify a different "
-                   "top-level directory to store science data.")))
+    print(''.join(("\nHi there!  Please inform pysat where you will store "
+                   "(or are storing) science data by "
+                   "running pysat.utils.set_data_dir and specifying "
+                   "a location.")))
 else:
     # load up stored data path
     with open(os.path.join(pysat_dir, 'data_path.txt'), 'r') as f: